refactor: test/config 폴더명 변경
