+++ src/test/java/org/ahpuh/surf/user/controller/UserControllerTest.java
@@ -1,7 +1,7 @@
 package org.ahpuh.surf.user.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.ahpuh.surf.mockaws.MockAwsS3Service;
+import org.ahpuh.surf.config.MockAwsS3Service;
 import org.ahpuh.surf.user.dto.UserJoinRequestDto;
 import org.ahpuh.surf.user.dto.UserLoginRequestDto;
 import org.ahpuh.surf.user.dto.UserUpdateRequestDto;
